Added add item functionality (not working) and RecyclerView ListActivity

+++ app/src/main/java/com/zerrium/uts/SqliteHelper.java
@@ -4,18 +4,16 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteException;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.os.Build;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 
 import java.util.ArrayList;
 
 public class SqliteHelper extends SQLiteOpenHelper {
 
-    public static final String DATABASE_NAME = "itemdb";
+    public static final String DATABASE_NAME = "item.db";
     public static final int DATABASE_VERSION = 1;
     private static final String TABLE_ITEMS = "items";
     private static final String KEY_ID = "id";
@@ -23,14 +21,13 @@ public class SqliteHelper extends SQLiteOpenHelper {
     private static final String KEY_QTY = "qty";
     private static final String KEY_DESC = "descr";
     private static final String SQL_TABLE_ITEMS = "CREATE TABLE " + TABLE_ITEMS +
-            "( "+ KEY_ID +" INTEGER PRIMARY KEY, " +
+            "( "+ KEY_ID +" TEXT PRIMARY KEY, " +
             KEY_NAME +" TEXT, " +
             KEY_QTY + " INTEGER, " +
             KEY_DESC + " TEXT)";
 
-    @RequiresApi(api = Build.VERSION_CODES.P)
-    public SqliteHelper(@Nullable Context context, @Nullable String name, int version, @NonNull SQLiteDatabase.OpenParams openParams) {
-        super(context, name, version, openParams);
+    public SqliteHelper(@Nullable Context context) {
+        super(context, DATABASE_NAME, null, DATABASE_VERSION);
     }
 
     @Override
@@ -44,47 +41,53 @@ public class SqliteHelper extends SQLiteOpenHelper {
     }
 
     //CRUD methods (create, read, update, delete)
-    protected void addItem(Item item){
+    protected void addItem(Item item) throws SQLiteException{
+        if (item == null) return;
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues v = new ContentValues();
         v.put(KEY_ID, item.getId());
         v.put(KEY_NAME, item.getName());
         v.put(KEY_QTY, item.getQty());
         v.put(KEY_DESC, item.getDesc());
-        long todo_id = db.insert(TABLE_ITEMS, null, v);
+        db.insert(TABLE_ITEMS, null, v);
     }
 
-    protected ArrayList<Item> getItem(){
-        SQLiteDatabase db = this.getReadableDatabase();
+    protected ArrayList<Item> getItem() throws SQLiteException{
+        SQLiteDatabase db = SqliteHelper.this.getReadableDatabase();
 
         Cursor cursor = db.query(
-                TABLE_ITEMS, //table
-                new String[]{KEY_ID, KEY_NAME, KEY_QTY, KEY_DESC}, //column
-                "*", //selection
-                null, //where
-                null, //groupby
-                null, //having
-                null); //orderby
+                    TABLE_ITEMS, //table
+                    new String[]{KEY_ID, KEY_NAME, KEY_QTY, KEY_DESC}, //column
+                    null, //selection
+                    null, //where
+                    null, //groupby
+                    null, //having
+                    null); //orderby
 
         ArrayList<Item> i = new ArrayList<>();
+
         while(cursor.moveToNext()){
             i.add(new Item(cursor.getString(0), cursor.getString(1), cursor.getInt(2), cursor.getString(3)));
         }
+
         cursor.close();
+
         return i;
     }
 
-    protected void updateItem(Item item){
+    protected void updateItem(Item item) throws SQLiteException {
+        if (item == null) return;
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues v = new ContentValues();
         v.put(KEY_NAME, item.getName());
         v.put(KEY_QTY, item.getQty());
         v.put(KEY_DESC, item.getDesc());
-        long todo_id = db.update(TABLE_ITEMS, v, KEY_ID + "LIKE ?", new String[]{item.getId()});
+        db.update(TABLE_ITEMS, v, KEY_ID + "LIKE ?", new String[]{item.getId()});
     }
 
-    protected void deleteItem(Item item){
+    protected void deleteItem(Item item) throws SQLiteException{
+        if (item == null) return;
         SQLiteDatabase db = this.getWritableDatabase();
-        long todo_id = db.delete(TABLE_ITEMS, KEY_ID + "LIKE ?", new String[]{item.getId()});
+        db.delete(TABLE_ITEMS, KEY_ID + "LIKE ?", new String[]{item.getId()});
     }
 }
